First validation concept for UI validation:
Created a "ValidationServices" in the app.js to be injected into the controller which has to validate the user entries. Validation adds "validationFail" and "validationMessage" to the property item. This can be used in the UI to switch the property into "error mode" and display the validation fail message.

To be done: create a "type" validator, each for int | boolean | decimal | string and validate based on the current configuration. Once this is working we can add a property called "validation" to the properties to do for example URL validation or whatever else is needed.

Validation also needs to be added to the proeprties of the test run and all the other edit fields in the BM server UI framework :-)